Updated svg-pan-zoom.js with a mouseHandle option
This new option serves as a reference to an element within the svg (or even the svg canvas itself) that acts as the intended recipient of the mouse interactions. Any interactions outside of this element will be completely ignored. Useful if you're using <foreignObject>s within the svg and <input>s within that, and you want to be able to mouse-interact with them.